Remove broken makefiles (#17)
PR #6 broke all the makefiles, which is confusing to users who try to run `make` (@yushangdi and @jeshi96 can attest to this). Let's remove them for now since they don't work. The plan is for @jeshi96 to follow up on this by writing makefiles that work. For the time being, we aim to maintain both Bazel and Make for this repository -- although Bazel is our primary build system, Julian requested that we keep an alternative build system for those who do not wish to install Bazel. 

## Changes
- Removed a lot of makefiles
- Removed a lot of `.gitignore` files whose only purpose was to ignore binaries created by makefiles
- Added BUILD files for `KTruss` and `utils/`. Did not attempt running all the binaries. `KTruss_main` segfaults; `utils:compressor` seems to work on Ubuntu with g++.
- Added BUILD files for  `Connectivity/Framework` and `Connectivity/Incremental`. @ldhulipala should confirm that they capture the intent of the original makefiles. I did not attempt running the binaries.
  - Adding these BUILD targets seems to have doubled compilation time. Perhaps refactoring could mitigate this -- I see places where we use a lot of templating on enums when it might be possible that we can instead give the enum as an argument and use a switch statement instead.