scripts: Add script for running all benchmarks (#25)
This PR adds a script for running all benchmarks. Run it using a command like
```sh
python3.7 run_all_benchmarks.py --unweighted_graph=~/gbbs/inputs/rMatGraph_J_5_100 --weighted_graph=~/gbbs/inputs/rMatGraph_WJ_5_100 --timeout=60
```
It needs Python >=3.7.

The script hardcodes lists of benchmarks that should be run on an unweighted graph vs. on a weighted graph. The script checks that all benchmarks are accounted for so that no benchmarks will slip through the cracks.

We could put this into CI as an integration test, although some additional logic needs to be put into CI -- right now I'm guessing that everything would fail to run due to `pbbslib/list_allocator.h`'s `default_alloc_size` being too large for the CI machine to handle.

What this script does _not_ do:
  - Does not check correctness of benchmark outputs. The script only checks that the benchmarks run and exit with error code 0.
  - Does not run benchmarks on asymmetric graphs.
  - Does not run "//benchmarks/Connectivity/Incremental/mains:*_no_starting" benchmarks --- their input format is a bit different than that of other benchmarks, and I'm not sure how to run them.

We could modify the script to be able to run on asymmetric graphs, but I've ignored that for this initial version of the script for simplicity.