SCAN: reduce contention on atomic counters
This is an intrusive optimization (in the sense that we have to modify
Triangle.h). It reduces the shared neighbor computation time on an AWS
c5.4xlarge instance on the orkut graph from ~23.5 to ~22.5 sec (5%
improvement) (total index construction time is ~30 sec). Is it worth it?
Committing for now, but consider reverting. It appears that this may
also slow down the triangle counting benchmark slightly (~0.5%).