SCAN: faster neighbor order construction
boy oh boy is this gross.

bottleneck was that we were accessing sparse_tables a whole bunch of
times which is real slow. replace the sparse table with an array and
things get a lot faster. however, this requires some gross interface
changes, which I dealt with by copying code from Triangle.h and vertex.h