seq: Give operator[]  const and non-const overloads (#26)
`pbbs::range` and `pbbs::sequence` have a `T& operator[](size_t) const` operator, which has the odd effect that you can write to indices in a const `range`/`sequence`. 

This PR redefines `operator[]` on `pbbs::range` and `pbbs::sequence` so that you can write to indices in a non-`const` `range`/`sequence` but you can only read indices in a `const` `range`/`sequence`.

This should be a no-op in terms of runtime behavior.