benchmarks: Crudely fix Makefile link order issues
Wrap all linked objects in --start-group and --end-group. The `ld`
manual says "Using this option has a significant performance cost. It is
best to use it only when there are unavoidable circular references
between two or more archives." However, we're going for simplicity
with these makefiles since it's not our main build system, so whatever