compressor: Fix usage of bytepd_amortized encoding
The compressor wasn't working with how the rest of GBBS handles
compressed graphs. My guess as to why this was happening: we
rearranged the directory structure of GBBS a while ago, which made the
compressor not compile. Commit 66de5dd885ceb0dcc1360d97269b18dcada179d4
fixed the compilation errors by following the compiler suggestion to add
a `byte::` namespace to some encoding functions that compressor.cc uses,
but the correct namespace to use is actually `bytepd_amortized::`.

This commit fixes the namespacing and makes the compressor work
properly.

Testing:
- Confirm that the compressor builds with both Bazel and Make.
- Run the compressor on the SNAP dblp graph.
- Run `benchmarks/BFS/NonDeterministicBFS:BFS_main` on the compressed
dblp graph with arguments `-s -c -src 0` and confirm that this no longer
segfaults.